requirements.txt
dj-database-url==1.2.0
Django==4.1.5
django-admin-interface
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.8.6

#For Static Files
whitenoise==6.2.0
Brotli==1.0.9
